refactor: extract datago envelope parser (#267)
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
